Restructure presenter folders
