+++ tweeter-web/src/presenter/mainLayout/FolloweePresenter.ts
@@ -1,5 +1,5 @@
 import { AuthToken } from "tweeter-shared";
-import { FollowService } from "../model/service/FollowService";
+import { FollowService } from "../../model/service/FollowService";
 import { UserItemPresenter, UserItemView } from "./UserItemPresenter";
 
 export const PAGE_SIZE = 10;
