+++ tweeter-web/src/presenter/mainLayout/StoryPresenter.ts
@@ -1,10 +1,10 @@
 import { AuthToken } from "tweeter-shared";
-import { StatusService } from "../model/service/StatusService";
+import { StatusService } from "../../model/service/StatusService";
 import { StatusItemPresenter, StatusItemView } from "./StatusItemPresenter";
 
 export const PAGE_SIZE = 10;
 
-export class FeedPresenter extends StatusItemPresenter {
+export class StoryPresenter extends StatusItemPresenter {
   private _statusService: StatusService;
 
   constructor(view: StatusItemView) {
@@ -14,7 +14,7 @@ export class FeedPresenter extends StatusItemPresenter {
 
   public async loadMoreItems(authToken: AuthToken, userAlias: string) {
     try {
-      const [newItems, hasMore] = await this._statusService.loadMoreFeedItems(
+      const [newItems, hasMore] = await this._statusService.loadMoreStoryItems(
         authToken,
         userAlias,
         PAGE_SIZE,
@@ -26,7 +26,7 @@ export class FeedPresenter extends StatusItemPresenter {
       this.view.addItems(newItems);
     } catch (error) {
       this.view.displayErrorMessage(
-        `Failed to load feed because of exception: ${error}`
+        `Failed to load story because of exception: ${error}`
       );
     }
   };
